# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/v1/users/{iid}:
    x-swagger-router-controller: user
    get:
      description: Returns all users
      # used as the method name of the controller
      operationId: user
      parameters:
        # MANDATORY HEADERS
        - name: iid
          in: path
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      description: update users
      # used as the method name of the controller
      operationId: updateUser
      parameters:
        # MANDATORY HEADERS
        - name: iid
          in: path
          type: string
          required: true

        # BODY PARAMS
        - name: userBody1
          in: body
          schema:
            $ref: "#/definitions/UserBody1"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: Delete user
      # used as the method name of the controller
      operationId: deleteUser
      parameters:
        # MANDATORY HEADERS
        - name: iid
          in: path
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/users/updateManyUsers:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    put:
      description: update users
      # used as the method name of the controller
      operationId: updateMultipleUsers
      parameters:
        # MANDATORY HEADERS
        - name: age
          in: query
          type: string
          required: false
        - name: salary
          in: query
          type: number
          required: false

        # BODY PARAMS
        - name: userBody1
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserBody2"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/users:
    x-swagger-router-controller: user
    post:
      description: add users
      # used as the method name of the controller
      operationId: addUser
      parameters:
        # BODY PARAMS
        - name: userBody
          in: body
          schema:
            $ref: "#/definitions/UserBody"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/users/find:
    x-swagger-router-controller: user
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: userFind
      parameters:
        - name: city
          in: header
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  UserBody:
    type: object
    properties:
      iid: 
        type: number
      name:
        type: string
      age:
        type: number
      city:
        type: string
      contact: 
        type: number
    required:
      - iid
      - name
      - age
      - city
      - contact
  UserBody1:
    type: object
    properties:
      name:
        type: string
      age:
        type: number
      city:
        type: string
      contact: 
        type: number
  UserBody2:
    type: object
    properties:
      name:
        type: string
      age:
        type: number
      city:
        type: string
      contact: 
        type: number
      company:
        type: string
      bonus:
        type: number
...